Position trophy button
