+++ browser/src/Services/Sidebar/SidebarContentSplit.tsx
@@ -127,6 +127,7 @@ export class SidebarHeaderView extends React.PureComponent<ISidebarHeaderProps, 
 export const SidebarInnerPaneWrapper = withProps<{}>(styled.div)`
     flex: 1 1 auto;
     overflow-y: auto;
+    position: relative;
 `
 
 export class SidebarContentView extends React.PureComponent<
