+++ browser/src/UI/components/common.ts
@@ -16,6 +16,28 @@ const {
     IThemeColors
 >
 
+export interface ContainerProps {
+    direction: "horizontal" | "vertical"
+    fullHeight?: boolean
+    fullWidth?: boolean
+}
+
+export const Fixed = styled.div`
+    flex: 0 0 auto;
+`
+
+export const Full = styled.div`
+    flex: 1 1 auto;
+`
+
+export const Container = withProps<ContainerProps>(styled.div)`
+    display: flex;
+    flex-direction: ${p => (p.direction === "vertical" ? "column" : "row")};
+
+    ${p => (p.fullHeight ? "height: 100%;" : "")}
+    ${p => (p.fullWidth ? "width: 100%;" : "")}
+`
+
 export type StyledFunction<T> = styledComponents.ThemedStyledFunction<T, IThemeColors>
 
 export function withProps<T, U extends HTMLElement = HTMLElement>(
